New upstream version 4.0.12

+++ jstests/aggregation/bugs/server3253.js
@@ -1,3 +1,10 @@
+// Cannot implicitly shard accessed collections because unsupported use of sharded collection
+// for output collection of aggregation pipeline.
+// @tags: [
+//   assumes_superuser_permissions,
+//   assumes_unsharded_collection,
+// ]
+
 // server-3253 Unsharded support for $out
 load('jstests/aggregation/extras/utils.js');
 
@@ -43,7 +50,7 @@ function test(input, pipeline, expected) {
 
 function listCollections(name) {
     var collectionInfosCursor = db.runCommand("listCollections", {filter: {name: name}});
-    return new DBCommandCursor(db.getMongo(), collectionInfosCursor).toArray();
+    return new DBCommandCursor(db, collectionInfosCursor).toArray();
 }
 
 input.insert({_id: 1});
@@ -58,7 +65,7 @@ assert.eq([], listCollections(/tmp\.agg_out/));
 
 // basic test
 test(input,
-         [{$project: {a: {$add: ['$_id', '$_id']}}}],
+     [{$project: {a: {$add: ['$_id', '$_id']}}}],
      [{_id: 1, a: 2}, {_id: 2, a: 4}, {_id: 3, a: 6}]);
 
 // test with indexes
@@ -66,7 +73,7 @@ assert.eq(output.getIndexes().length, 1);
 output.ensureIndex({a: 1});
 assert.eq(output.getIndexes().length, 2);
 test(input,
-         [{$project: {a: {$multiply: ['$_id', '$_id']}}}],
+     [{$project: {a: {$multiply: ['$_id', '$_id']}}}],
      [{_id: 1, a: 1}, {_id: 2, a: 4}, {_id: 3, a: 9}]);
 
 // test with empty result set and make sure old result is gone, but indexes remain
@@ -81,18 +88,18 @@ test(input, [{$project: {b: "$_id"}}], [{_id: 1, b: 1}, {_id: 2, b: 2}, {_id: 3,
 // test with full text index
 output.ensureIndex({c: "text"});
 assert.eq(output.getIndexes().length, 4);
-test(input,
-         [{$project: {c: {$concat: ["hello there ", "_id"]}}}],
-     [
-       {_id: 1, c: "hello there _id"},
-       {_id: 2, c: "hello there _id"},
-       {_id: 3, c: "hello there _id"}
-     ]);
+test(input, [{$project: {c: {$concat: ["hello there ", "_id"]}}}], [
+    {_id: 1, c: "hello there _id"},
+    {_id: 2, c: "hello there _id"},
+    {_id: 3, c: "hello there _id"}
+]);
 
-// test with capped collection
-cappedOutput.drop();
-db.createCollection(cappedOutput.getName(), {capped: true, size: 2});
-assertErrorCode(input, {$out: cappedOutput.getName()}, 17152);
+// test with capped collection, skip for mobile SE as it doesn't support capped collections.
+if (jsTest.options().storageEngine !== "mobile") {
+    cappedOutput.drop();
+    db.createCollection(cappedOutput.getName(), {capped: true, size: 2});
+    assertErrorCode(input, {$out: cappedOutput.getName()}, 17152);
+}
 
 // ensure everything works even if input doesn't exist.
 test(inputDoesntExist, [], []);
